Move loadBooks function to functions.js file
